B2XKlaim-BackEnd/src/test/java/com/example/B2XKlaim/ServiceTest/Parser/BpmnElementFactoryTest.java
package com.example.B2XKlaim.ServiceTest.Parser;

import com.example.B2XKlaim.Service.Parser.BpmnElementFactory;
import com.example.B2XKlaim.Service.bpmnElements.BpmnElement;
import com.example.B2XKlaim.Service.bpmnElements.objects.pool.PL;
import org.camunda.bpm.model.bpmn.instance.MessageFlow;
import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;
import org.w3c.dom.Document;
import org.w3c.dom.Element;
import org.w3c.dom.NodeList;

import static org.junit.jupiter.api.Assertions.*;

public class BpmnElementFactoryTest {

    private Document document;
    private BpmnElementFactory factory;

    @BeforeEach
    public void setUp() throws Exception {
        // TODO: Initialize 'document' with a BPMN XML
        // This can be done using a mock library or manually parsing a sample BPMN XML string.
        factory = new BpmnElementFactory(document);
    }

    @Test
    public void testCreateMessageFlowElement() throws Exception {
        // TODO: Create an Element instance representing a 'bpmn:messageFlow'
        NodeList elementNodes = document.getElementsByTagName("*");
        Element element = (Element) elementNodes.item(0);
        BpmnElement result = factory.createBpmnElement(element);

        assertTrue(result instanceof MessageFlow, "Expected a MessageFlow instance");
        // TODO: Add more assertions to check the properties of the returned MessageFlow
    }

    @Test
    public void testCreateParticipantElement() throws Exception {
        // TODO: Create an Element instance representing a 'bpmn:participant'
        NodeList elementNodes = document.getElementsByTagName("*");
        Element element = (Element) elementNodes.item(0);
        BpmnElement result = factory.createBpmnElement(element);

        assertTrue(result instanceof PL, "Expected a PL instance");
        // TODO: Add more assertions to check the properties of the returned PL
    }

    // Add more tests for other BPMN elements...
}